Update sidebars.js
In this updated code, a delay function is added to the script which delays the execution of the code by 5 seconds using a promise. After the 5 seconds have passed, the sidebars object is exported. This will increase the time it takes for the script to execute, making it more "expensive" in terms of resource usage.